wire code to use new store + fix build
