+++ packages/webapp/src/store/update.ts
@@ -7,22 +7,15 @@
 // =================================================================================================
 
 import { getDefaultStore } from "jotai"
-import { getAccount, getNetwork, readContract, watchAccount, watchNetwork } from "wagmi/actions"
+import { getAccount, getNetwork, watchAccount, watchNetwork } from "wagmi/actions"
 
-import { deployment } from "src/deployment"
-import { gameABI } from "src/generated"
-import {
-  gameID,
-  gameStatus,
-  hasVisitedBoard,
-  isGameCreator,
-  isGameJoiner
-} from "src/store"
-import { gameCards_, gameData_, playerAddress_, randomness_ } from "src/store/private"
 import { subscribeToGame } from "src/store/subscriptions"
 import { GameStatus, type FetchedGameData, currentPlayerAddress, getGameStatus } from "src/types"
 import { AccountResult, chains, NetworkResult } from "src/chain"
 import { formatTimestamp } from "src/utils/js-utils"
+import * as atoms from "src/store/atoms"
+import { Address } from "wagmi"
+import * as net from "src/store/network"
 
 // =================================================================================================
 // INITIALIZATION
@@ -51,40 +44,48 @@ export function setupStore() {
 
   // The game ID can change from actions in this tab, but also in other tabs, or can be retrieved
   // from the storage upon boot, so we need to listen to the storage.
-  store.sub(gameID, () => {
-    setDependencies(store.get(gameID))
+  store.sub(atoms.gameID, () => {
+    gameIDListener(store.get(atoms.gameID))
   })
 
   // Make sure we don't miss the initial value, if already set.
-  if (store.get(gameID) !== null)
-    setDependencies(store.get(gameID))
+  const gameID = store.get(atoms.gameID)
+  if (gameID !== null)
+    gameIDListener(gameID)
 }
 
 // =================================================================================================
-// CONSISTENT UPDATES
+// CHANGE LISTENERS
 
 /**
  * Called whenever the connected wallet address changes. Makes sure to clear the address if the
  * wallet is disconnected, as well as the game data.
  */
 function updatePlayerAddress(result: AccountResult) {
-  // normalize undefined to null
-  const oldAddress = store.get(playerAddress_) || null
+  const oldAddress = store.get(atoms.playerAddress)
+  // normalize to null (never undefined)
   const newAddress = result.status === 'disconnected' ? null : (result.address || null)
 
   if (oldAddress !== newAddress && isNetworkValid()) {
     console.log(`player address changed from ${oldAddress} to ${newAddress}`)
-    store.set(playerAddress_, newAddress)
+    store.set(atoms.playerAddress, newAddress)
 
-    // TODO temporary hack: don't reset the game data on hard page refresh, because the player address will transition from null to the connected address
+    // TODO the below should go away if we stop saving the gameID in browser storage and just fetch
+    //      the gameID a player is in instead
+
+    // The check is important: on a page load when already connected, the address will transition
+    // from null to the connected address, and we don't want to throw away the game ID.
+    //
+    // If the old address is null, there is also nothing to reset (excepted at load).
     if (oldAddress !== null)
-      store.set(gameID, null) // resets all game data
+      // New player means current game & game data is stale, reset it.
+      store.set(atoms.gameID, null)
   }
 }
 
 // -------------------------------------------------------------------------------------------------
 
-/** Returns true if the network we are connected to is the one we support ({@link chain}). */
+/** Returns true if the network we are connected to is the one we support ({@link chains}). */
 function isNetworkValid(network: NetworkResult = getNetwork()) {
   return chains.some(chain => chain.id === network.chain?.id)
 }
@@ -98,7 +99,7 @@ function isNetworkValid(network: NetworkResult = getNetwork()) {
 function updateNetwork(result: NetworkResult) {
   console.log(`network changed to chain with ID ${result.chain?.id}`)
   if (!isNetworkValid(result)) {
-    store.set(gameID, null) // resets all game data
+    store.set(atoms.gameID, null) // resets all game data
   }
 }
 
@@ -116,14 +117,13 @@ function updateNetwork(result: NetworkResult) {
  * It never causes race conditions or weird data states: this resets all associated states, and if
  * a refresh lands with another ID, it will be ignored.
  */
-function setDependencies(ID: bigint|null) {
+function gameIDListener(ID: bigint|null) {
   console.log(`transitioning to game ID ${ID}`)
 
   // avoid using inconsistent data
-  store.set(gameData_, null as FetchedGameData)
-  store.set(hasVisitedBoard, false)
-
-  // TODO reset per-player state here, when we start using it
+  store.set(atoms.gameData, null as FetchedGameData)
+  store.set(atoms.hasVisitedBoard, false)
+  store.set(atoms.randomness, null)
 
   subscribeToGame(ID) // will unusubscribe if ID is null
   if (ID !== null)
@@ -132,160 +132,147 @@ function setDependencies(ID: bigint|null) {
 }
 
 // =================================================================================================
+// REFRESH GAME DATA
 
-// Used to throttle refreshes: at most one refresh can be requested in flight, unless two seconds
-// elapse. After a refresh lands, another one become spossible immediately.
-const refreshThrottle = 2000
-let lastRequestTimestamp = 0
-
-// Used to avoid "zombie" refreshes: old refreshes overwriting newer game data.
-let sequenceNumber = 1
-let lastCompletedNumber = 0
+// -------------------------------------------------------------------------------------------------
 
 /** If the game has started and we don't have the cards yet, we need to fetch them. */
 function shouldUpdateCards(): boolean {
-  return store.get(gameStatus) >= GameStatus.STARTED && store.get(gameCards_) === null
+  return store.get(atoms.gameStatus) >= GameStatus.STARTED && store.get(atoms.gameCards) === null
 }
 
+// -------------------------------------------------------------------------------------------------
+
 /**
- * Triggers a manual refresh of the game data, setting the {@link gameData} atom.
- * If the game ID changes the while the refresh is in flight, the refresh is ignored.
+ * Returns whether an update for the given gameID and player address is stale, i.e. if the current
+ * store gameID and player address are different, meaning they change underneath the fetch and the
+ * fetched data should be discarded.
+ */
+function isStale(ID: bigint, player: Address): boolean {
+  const storeID = store.get(atoms.gameID)
+  const storePlayer = store.get(atoms.playerAddress)
+  if (player !== storePlayer) {
+    console.log(`Rejected stale data with player ${player} (current: ${storePlayer})`)
+    return true
+  }
+  if (ID !== storeID) {
+    console.log(`Rejected stale data with game ID ${ID} (current: ${storeID})`)
+    return true
+  }
+  return false
+}
+
+// -------------------------------------------------------------------------------------------------
+
+/**
+ * Fetches the game data or the given game ID and player address, and updates the store with the
+ * fetched data (if not throttled, zombie, or stale). Returns the fetched data.
+ */
+async function updateGameData(ID: bigint, player: Address): Promise<FetchedGameData> {
+  const gameData = await net.fetchGameData(ID)
+  if (gameData === null || isStale(ID, player)) return
+
+  const timestamp = Date.now()
+  console.groupCollapsed(`updated game data (at ${formatTimestamp(timestamp)})`)
+  console.dir(gameData)
+  console.groupEnd()
+
+  const oldGameData = store.get(atoms.gameData)
+  if (oldGameData !== null && oldGameData.lastBlockNum >= gameData.lastBlockNum)
+    // We already have more or as recent data, no need to trigger a store update.
+    return oldGameData
+
+  store.set(atoms.gameData, gameData)
+  void updateRandomness(ID, player, gameData)
+  return gameData
+}
+
+// -------------------------------------------------------------------------------------------------
+
+/**
+ * Fetches the game cards for the given game ID and player address (if not throttled, zombie, or
+ * stale). Returns the fetched cards.
+ */
+async function fetchCards(ID: bigint, player: Address): Promise<readonly bigint[]> {
+  const cards = await net.fetchCards(ID)
+  if (cards === null || isStale(ID, player)) return
+  return cards
+}
+
+// -------------------------------------------------------------------------------------------------
+
+/**
+ * Fetches the randomness for the given game ID, player address and block number (if not throttled,
+ * zombie, or stale) and updates the store accordingly.
+ */
+async function updateRandomness(ID: bigint, player: Address, gameData: FetchedGameData) {
+
+  // (1) The current player is not us, no need to fetch randomness.
+  // (2) The randomness is meaningless before the game starts.
+  //     (This should be covered by the player check, but let's be safe.)
+  if (player !== currentPlayerAddress(gameData)
+  ||  getGameStatus(gameData, player) < GameStatus.STARTED)
+  {
+    store.set(atoms.randomness, null) // not strictly necessary, doesn't hurt to be safe
+    return
+  }
+
+  // At this point, we know the gameData (hence its lastBlockNum) changed, and that the current
+  // player is us, so we need the randomness.
+
+  // Randomness is needed, but we do not have it yet. The previous value is incorrect, erase it.
+  store.set(atoms.randomness, null)
+
+  const randomness = await net.fetchRandomness(gameData.lastBlockNum)
+  if (randomness === null || isStale(ID, player)) return
+
+  const timestamp = Date.now()
+  console.log(`updated randomness (at ${formatTimestamp(timestamp)})`)
+
+  store.set(atoms.randomness, randomness)
+}
+
+// -------------------------------------------------------------------------------------------------
+
+/**
+ * Triggers a refresh of the game data, setting the {@link atoms.gameData} atom. If the game ID or
+ * the player changes the while the refresh is in flight, the refresh is ignored.
+ *
+ * If necessary ({@link shouldUpdateCards} returns true), also fetches the cards and updates the
+ * {@link atoms.gameCards} atom accordingly. Similarly updates the randomness {@link
+ * atoms.randomness} if needed.
  *
  * @param forceFetchCards forces fetching the cards even though {@link shouldUpdateCards} initially
  * returns false. This is useful when we know that the new game data will move us to a state where
- * we should update the cards. We still check if the update is necessary anyhow.
+ * we should update the cards.
  */
 export async function refreshGameData({ forceFetchCards = false } = {}) {
-  const ID = store.get(gameID)
-  const player = store.get(playerAddress_)
+  const ID = store.get(atoms.gameID)
+  const player = store.get(atoms.playerAddress)
 
   if (ID === null) {
     console.error("refreshGameData called with null ID")
     return
   }
 
-  // === Throttle ===
+  const shouldFetchCards = shouldUpdateCards() || forceFetchCards
 
-  const seqNum = sequenceNumber++
-  const timestamp = Date.now()
-  if (timestamp - lastRequestTimestamp < refreshThrottle)
-    return // there is a recent-ish refresh in flight
-  lastRequestTimestamp = timestamp
-
-  // === Fetch ===
-
-  const gameDataPromise = readContract({
-    address: deployment.Game,
-    abi: gameABI,
-    functionName: "fetchGameData",
-    args: [ID]
-  })
-
-  let gameCardsPromise: Promise<readonly bigint[]> = null
-
-  // If the game has started and we haven't got the cards yet, fetch them.
-  if (forceFetchCards || shouldUpdateCards()) {
-    gameCardsPromise = readContract({
-      address: deployment.Game,
-      abi: gameABI,
-      functionName: "getCards",
-      args: [ID]
-    })
+  if (!shouldFetchCards) {
+    void updateGameData(ID, player)
+  } else {
+    // We need the game data in order to update the cards in the store.
+    const gameDataPromise = updateGameData(ID, player)
+    const cardsPromise = fetchCards(ID, player)
+    const gameData = await gameDataPromise
+    const cards = await cardsPromise
+    if (shouldUpdateCards()) {
+      // Don't (re)assign if cards were fetched in the meantime or if the game isn't started yet.
+      const decks = gameData.playerData.map(pdata => cards.slice(pdata.deckStart, pdata.deckEnd))
+      store.set(atoms.gameCards, {gameID: ID, cards, decks})
+      const timestamp = Date.now()
+      console.log(`updated cards (at ${formatTimestamp(timestamp)})`)
+    }
   }
-
-  // Await after firing all calls
-  const gameData = await gameDataPromise
-  const gameCards = await gameCardsPromise
-
-  // === Stale Update Filtering ===
-
-  if (seqNum < lastCompletedNumber) return // ignore zombie refresh
-  lastCompletedNumber = seqNum
-
-  const storeID = store.get(gameID)
-  const storePlayer = store.get(playerAddress_)
-  const stale = ID !== storeID || player !== storePlayer
-  if (stale) {
-    // The game ID changed underneath this update, ignore it.
-    console.log(`Rejected stale data with game ID ${ID} (current game ID is ${storeID})`)
-    return
-  }
-
-  // === Log Updates (1/2) ===
-
-  console.groupCollapsed(`fetched data (at ${formatTimestamp(timestamp)})`)
-  console.dir(gameData)
-  console.groupEnd()
-
-  // Allow another refresh immediately.
-  lastRequestTimestamp = 0
-
-  // === Update Store ===
-
-  // TODO only update if things changed?
-
-  store.set(gameData_, gameData)
-  const updateCards = shouldUpdateCards()
-
-  if (gameCards && updateCards) {
-    const decks = gameData.playerData.map(pdata =>
-      gameCards.slice(pdata.deckStart, pdata.deckEnd))
-    store.set(gameCards_, {
-      gameID: ID,
-      cards: gameCards,
-      decks: decks
-    })
-  }
-
-  // === Log Updates (2/2) ===
-
-  if (gameCards && updateCards) console.log("fetched game cards")
-  console.groupEnd()
-
-  // === Sanity Check ===
-  // NOTE: We don't yet support spectating games.
-
-  if(!store.get(isGameCreator) && !store.get(isGameJoiner)) {
-    console.warn(`Tracking a game (${store.get(gameID)}) we are not participating in.`)
-    // NOTE(norswap): hunting an heisenbug that I've seen happen after creating a game
-    console.log("address", store.get(playerAddress_))
-    console.log("creator", store.get(gameData_)?.gameCreator)
-  }
-
-  // === Fetch Randomness ===
-
-  // TODO
-
-  // The current player is not us, no need to fetch randomness.
-  if (player !== currentPlayerAddress(gameData))
-    return
-
-  // The randomness is meaningless before the game starts.
-  // This should be covered by the player check, but let's be safe.
-  if (getGameStatus(gameData, player) < GameStatus.STARTED)
-    return
-
-
-  // TODO maybe we already have the randomness?
-  //      e.g. gamedata did not change (should be tackled upstream)
-  //      or data changed, but player and lastBlockNum did not (I think impossible rn but should handle to be future-proof)
-
-  // Randomness is needed, but we do not have it yet. The previous value is incorrect, erase it.
-  store.set(randomness_, null)
-
-  const randomness = readContract({
-    address: deployment.Game,
-    abi: gameABI,
-    functionName: "getRandomness",
-    args: [gameData.lastBlockNum]
-  })
-
-  const randomNum = BigInt(await randomness)
-
-  // TODO stale filtering: ID, player, lastBlockNum
-
-  console.log("fetched randomness: ", randomNum)
-  store.set(randomness_, randomNum)
 }
 
 // =================================================================================================
